Initial Design Document
This is the first step of the coding challeng, providing a design document. I covered the requested areas,
and added some detail for some of the more interesting sections. Let me know if you'd like to see more code
in the design, such as examples of how certain features might look.